Refactor reorganize theme files and update file names
This commit reorganizes the theme files to create a more modular and maintainable structure. I've created subfolders for `palette`, `surfaceColors`, `shadows`, and `elevation`, and nested related files within them.

In addition to restructuring the files, I've updated some of the file names to better reflect their contents. For example, `createGradient.js` is now located in the `surfaceColors` folder and `constants.js` files have been added to the `shadows` and `elevation` modules.

These changes should make it easier to manage the theme files and update specific parts of the theme as needed.
