+++ california-dev/src/themes/ThemeContext.js
@@ -1,11 +1,11 @@
 import { ThemeProvider } from "styled-components";
-import { lightTheme } from "./light";
-import { darkTheme } from "./dark";
+import { lightTheme } from "./light/lightTheme";
+import { darkTheme } from "./dark/darkTheme";
 import { GlobalStyles, StylesReset } from "./GlobalStyles";
 
 
 
-export const Theme = ({ children, theme }) => {
+export const ThemeContext = ({ children, theme }) => {
   return (
     <ThemeProvider theme={theme === "light" ? lightTheme : darkTheme}>
       <GlobalStyles />
