+++ california-dev/src/themes/dark/darkTheme.js
@@ -3,5 +3,5 @@ import { elevation } from "./elevation";
 
 export const darkTheme = {
   palette,
-  elevation,
+  elevation
 };
