+++ california-dev/src/themes/dark/palette/index.js
@@ -1,0 +1,1 @@
+export * from "./palette";
