+++ california-dev/src/themes/index.js
@@ -1,1 +1,1 @@
-export * from "./Theme";
+export * from "./ThemeContext";
