+++ california-dev/src/themes/light/lightTheme.js
@@ -4,4 +4,5 @@ import { elevation } from "./elevation";
  export const lightTheme = {
   palette,
   elevation
+
 }
