+++ california-dev/src/themes/light/palette/index.js
@@ -1,0 +1,1 @@
+export * from "./palette";
